+++ gDRutils/tests/testthat/test-headers.R
@@ -1,22 +1,13 @@
 test_that("get_header works", {
-  reset_identifiers()
-  reset_headers()
+  reset_env_identifiers()
 
   expect_error(get_header("BOGUS"))
   expect_equal(get_header("manifest"), c("Barcode", "Template", "Duration"))
   expect_equal(length(get_header()), 12)
-})
 
+  set_env_identifier("duration", "TEST_DURATION")
+  expect_equal(get_header("manifest"), c("Barcode", "Template", "TEST_DURATION"))
 
-test_that("reset_headers works", {
-  reset_identifiers()
-  reset_headers()
-
-  set_identifier("duration", "TEST_DURATION")
-  expect_equal(get_header("manifest"), "TEST_DURATION")
-
-  reset_identifiers()
-  hlist <- reset_headers()
+  reset_env_identifiers()
   expect_equal(get_header("manifest"), c("Barcode", "Template", "Duration"))
-  expect_equal(hlist, NULL)
 })